+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2025-08-12
+
+- `.add` and `.edit` now react with a checkmark instead of sending the full success message.
+
 ## 2025-08-10
 
 **JS execution:**